requirements.txt
streamlit
langchain
openai
python-dotenv
pdfminer.six
tiktoken
langchain-openai